# backend/openapi/bundled.yaml
openapi: 3.0.0
info:
  title: Podverse Admin API
  version: 1.0.0
  description: Admin API for Podverse podcast platform dashboard
servers:
  - description: Local development
    url: http://localhost:8000/admin
security:
  - Auth0Bearer: []
paths:
...
  /channels:
    get:
      summary: Retrieve a list of channels
      description: |
        Returns a paginated list of channels with their basic metadata.
        Supports filtering by text-based search on `title`, and sorting by `id` or `title`.
        Primary table is `channel`.
        Related tables are :
        - `medium`: used to include the medium type (e.g. podcast, video) 
        - `feed`: used to show the associated RSS feed URL or status (optional).
      tags:
        - Channels
      security:
        - Auth0Bearer: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              - id
              - title
          description: |
            Field to sort results by. Supported fields: `id`, `title`.
        - name: sort_order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: Sort order (ascending or descending).
      responses:
        '200':
          description: List of channels returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Channel'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of matching channels
                      limit:
                        type: integer
                        description: Limit applied to query
                      offset:
                        type: integer
                        description: Offset applied to query
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /channels/{id}:
    get:
      summary: Retrieve a channel by ID
      description: |
        Returns details for a specific channel.
        Includes associated relationships and metadata such as slug, podcast index ID, feed reference, and medium.
        Primary table is `channel`.
        Related tables are :
        - `feed`: provides the associated feed URL and parsing status
        - `medium`: provides the media type of the channel
        - `channel_category` + `category`: maps the channel to its categories
        - `stats_aggregated_channel`: optionally used to show precomputed statistics
      tags:
        - Channels
      security:
        - Auth0Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the channel to retrieve
      responses:
        '200':
          description: Channel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /stats/channels:
    get:
      summary: Retrieve aggregated channel statistics
      description: |
        Returns aggregated stats for channels such as total listens or views.
        Sorted by highest monthly count (month_current_count DESC).
        Supports optional filtering by time window (daily, weekly, monthly) and search by channel title or ID.
        Data is sourced from the `channel` and `stats_aggregated_channel` tables.
      tags:
        - Statistics
      security:
        - Auth0Bearer: []
      parameters:
        - $ref: '#/components/parameters/time_filter'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: Aggregated channel statistics returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StatsChannel'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      limit:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    Auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Auth0
  parameters:
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Max number of results to return
    offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
      description: Number of results to skip for pagination
    search:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Search term to filter results by name or title
    time_filter:
      name: time_filter
      in: query
      description: Filter results by time window (e.g., daily, weekly, monthly)
      required: false
      schema:
        type: string
        enum:
          - daily
          - weekly
          - monthly
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Invalid request parameters
        code:
          type: string
          description: Error code for programmatic handling
          example: INVALID_PARAMETERS
        details:
          type: object
          description: Additional error details
          additionalProperties: true
          example:
            field: rss_url
            reason: Invalid URL format
      required:
        - error
        - code
    Channel:
      type: object
      properties:
        id:
          type: integer
          description: Unique numeric ID for the channel
          example: 201
        id_text:
          type: string
          maxLength: 15
          nullable: true
          description: Public-facing short ID for URLs
          example: ch_8a3f91
        slug:
          type: string
          maxLength: 100
          nullable: true
          description: URL-friendly slug
          example: the-daily-news
        feed_id:
          type: integer
          nullable: true
          description: ID of the associated feed
          example: 104
        podcast_index_id:
          type: integer
          nullable: true
          description: External Podcast Index identifier
          example: 872361
        podcast_guid:
          type: string
          format: uuid
          nullable: true
          description: '`<podcast:guid>` from the RSS feed'
          example: 1e8f3d7c-7fc0-4b70-90e1-991ed05bc4c4
        title:
          type: string
          maxLength: 255
          description: Title of the podcast or media channel
          example: The Daily News Podcast
        sortable_title:
          type: string
          maxLength: 255
          nullable: true
          description: Lowercased, normalized version of title for sorting
          example: daily news podcast
        medium_id:
          type: integer
          nullable: true
          description: Foreign key to the `medium` table
          example: 2
        has_podcast_index_value:
          type: boolean
          description: Indicates if channel uses Podcast Index's value tag data
          example: false
        has_value_time_splits:
          type: boolean
          description: Indicates whether channel items use value time split logic
          example: true
      required:
        - id
        - title
...
    StatsChannel:
      type: object
      properties:
        channel_id:
          type: integer
          description: Unique ID of the channel
        podcast_index_id:
          type: integer
          description: Podcast Index ID
        title:
          type: string
          description: Title of the podcast show
        total_listens:
          type: integer
          description: Total listens during the selected time period
        start_date:
          type: string
          format: date
          description: Start date of the aggregation window (ISO 8601)
        end_date:
          type: string
          format: date
          description: End date of the aggregation window (ISO 8601)
      required:
        - channel_id
        - podcast_index_id
        - title
        - total_listens
...
  responses:
    BadRequest:
      description: Bad request (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...